backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.1.0
openai==1.6.1
anthropic>=0.16.0
httpx==0.25.2
aiohttp==3.9.1
python-multipart==0.0.6
reportlab==4.0.7
pillow==10.1.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
tenacity==8.2.3
python-dateutil==2.8.2
pytz==2023.3
geopy==2.4.1
folium==0.15.1
beautifulsoup4==4.12.2
lxml==4.9.3
dedalus-labs

